Review: Minor refactoring
