+++ spring-hexagonal/src/main/java/com/baeldung/hexagonal/HexagonalApplication.java
@@ -5,9 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
 public class HexagonalApplication {
-
     public static void main(String[] args) {
         SpringApplication.run(HexagonalApplication.class, args);
     }
-
 }
