+++ spring-hexagonal/src/test/java/com/baeldung/hexagonal/HexagonalApplicationTests.java
@@ -1,4 +1,4 @@
-package com.baeldung.hexagonal.api;
+package com.baeldung.hexagonal;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
